refactor: 도메인에서는 UI를 사용하지 않도록 수정
- model 패키지를 view 패키지로 이동
